add benefits description and refactor types
