+++ src/graphql/types/types.js
@@ -1,0 +1,4 @@
+export { BonusInput, BonusType } from "./bonusType";
+export { LocationInput, LocationType } from "./locationType";
+export { OfferConnection, OfferType, CreateOfferInput } from "./offerType";
+export { StudentType, CreateStudentInput } from "./studentType";
